refactor(#97): lift isDirty to EditingContext; useDirtyFormBlock → src/hooks
Sets up the dirty-signal channel that #91's sort/pagination guard needs,
and lifts the (already generic) beforeunload hook out of the feature dir
so the next sidebar-edit consumer (vehicle-types, deck-plans) inherits
both pieces on day one — no further reaching into vehicles/ from chrome.

- EditingContext gains `isEditorDirty` + `setEditorDirty`.
- VehicleDetails pushes `isDirty` into the context via effect and clears
  on unmount so the signal doesn't leak across routes.
- `useDirtyFormBlock` git-mv'd from data/vehicles/hooks/ → src/hooks/;
  VehicleDetails and VehicleCreatePage imports updated.
